refactor : 패키지 구조, 패키지명, 변수명 수정
